fix: rework on custom newPlayer page #173
